Fix message state icon

+++ src/components/chat-page/account-editor/DeleteAccountSection.tsx
@@ -63,7 +63,7 @@ async function operateDeleteAccount(): Promise<void> {
         'first assign another user as the admin.',
       12.5,
     );
-  else {
+  else if (response !== undefined) {
     Storage.deleteTokenSet();
     await logOut();
   }
